% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binin.R
\name{binin}
\alias{binin}
\title{Spectral data binning}
\usage{
binin(X, ppm, width = NULL, npoints = NULL)
}
\arguments{
\item{X}{num matrix, NMR data with spectra in rows}

\item{ppm}{num array, chemical shift positions, length matches to columns in \code{X}}

\item{width}{num, bin size in ppm or NULL in case \code{npoints} is specified}

\item{npoints}{num, desired number of bins per spectrum or NULL in case  \code{width} is specified}
}
\value{
Numeric matrix with spectra in rows and chemical shift variables in columns.
}
\description{
Spectral data binning
}
\details{
Equidistant binning of spectra. Specify either \code{width} or \code{npoints} argument - if both are provided, \code{npoints} is used. Input argument \code{ppm} can be omitted if chemical shift information is encoded in the column names of the NMR matrix \code{X}.
}
\examples{
data(X, ppm)
Xb <- binin(X, ppm, width = 0.01)
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
