% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl.R
\name{bl}
\alias{bl}
\title{Baseline Correction}
\usage{
bl(X)
}
\arguments{
\item{X}{the spectrum to be baseline corrected}
}
\value{
an x with a corrected baseline
}
\description{
\code{bl()} removes the arched baseline from an NMR spectrum making the analysis more robust
}
\details{
NMR spectra often have broad signals produced by proteins within a sample that span a large portion of the ppm axis. These broad signals interfere with the analysis of specific peaks. By removing it, the anaylsis is more robust.
This is achieved by calculating the trend of the spectrum using \code{ptw::asysm()} and then subtracts that from the spectrum's values.
}
\examples{
read_in(path = system.file('extdata', package = 'concentr8r'),
        exp_type = list(exp=c("PROF_URINE_NOESY")),
        n_spec = 'multiple')
Xb <- bl(X)
}
\seealso{
Other {preproc}: 
\code{\link{preproc}()}
}
\author{
Kyle Bario \email{kylebario1@gmail.com}
}
\concept{{preproc}}
