% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cali.R
\name{cali}
\alias{cali}
\title{Chemical Shift Calibration}
\usage{
cali(X, ppm, type = c("tsp"))
}
\arguments{
\item{X}{num matrix, NMR data matrix with rows representing spectra and columns representing chemical shift position}

\item{ppm}{num vector, matched to columns of X}

\item{type}{str or num. Str: Either 'tsp' or 'glucose' for urine or blood-derived spectra, respectively (see Details). Num: ppm range of max height signal that will be used to reference to zero}
}
\value{
num matrix X, calibrated NMR data matrix.
}
\description{
Chemical shift calibration with a reference signal.
}
\details{
Spectral calibration to a selected chemical shift reference signal.
\code{type='tsp'}: calibration to 0 ppm using the highest peak located in interval 0 +/- 0.20 ppm (Trimethylsilylpropanoic acid resonance)
}
\examples{
read_in(path = system.file('extdata',package='concentr8r'),
        exp_type = list(exp=c("PROF_URINE_NOESY")),
        n_spec = 'multiple')
plot(ppm, X[2,], type = 'l', col = 'red',
    main = 'NMR Spectra (Processed vs. Non Processed)',
    xlab = 'Chemical Shift (ppm)', ylab = 'Intensity', xlim = c(10,-1))
Xc=cali(X, ppm, type='tsp')
points(ppm, X[2,], type = 'l', col = 'blue')
legend('topleft', legend = c("Uncalibrated", "Calibrated"), col = c('red', 'blue'), lty = 1)

}
\references{
Dona, A.C., \emph{et al.} (2014) Precision high-throughput proton NMR spectroscopy of human urine, serum, and plasma for large-scale metabolic phenotyping. \emph{Analytical Chemistry}. 86.19. 9887-94.
}
\author{
\email{torben.kimhofer@murdoch.edu.au}
}
