% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creNorm.R
\name{creNorm}
\alias{creNorm}
\title{Creatinine Normalisation}
\usage{
creNorm(X, ppm = NULL, cre3 = c(3, 3.1), cre4 = c(4, 4.1), err = 5)
}
\arguments{
\item{X}{The spectra intended to be normalised. Can either be a single spectrum in the form of a numerical array or multiple spectra in a numerical matrix with the rows being the spectra/samples and the columns being the ppm variables}

\item{ppm}{A numerical array holding the chemical shift values of the X matrix. Only necessary when X is an array, not when X is a matrix}

\item{cre3}{A concatenated numerical value of the lower and upper ppm values where the creatinine peak at 3.05 starts and ends.}

\item{cre4}{A concatenated numerical value of the lower and upper ppm values where the creatinine peak at 4.05 starts and ends.}

\item{err}{The level of error given when calculating the creatinine peak ratios. interperted as a percentage (i.e., 5 = 5\%)}
}
\value{
This function assigns the normalised X argument (as X_cre) and the calculated dilution factors (as dilf_cre) to the global environment.
}
\description{
Creatinine Normalisation (CN) is a useful method much like region of interest normalisation that can normalise spectra based on the total area of the creatinine signal at the chemical shift 3.05ppm.
}
\details{
\code{creNorm()} works by dividing each element in a row with the sum of the values from its Creatinine signal.
}
\examples{
# When X contains multiple spectra, ppm is not required
data(X)
creNorm(X)
cat(dilf_cre)

# When X has only one spectrum, ppm is required
data(X, ppm)
creNorm(X[1,], ppm)
cat(dilf_cre)

}
\seealso{
More on the methodology of CN and issue with using it are outlined here: \url{https://doi.org/10.1021/ac051632c}

Other {Attribute-Based}: 
\code{\link{q1Norm}()},
\code{\link{roiNorm}()},
\code{\link{taNorm}()},
\code{\link{vecNorm}()}
}
\author{
\email{kylebario1@gmail.com}
}
\concept{{Attribute-Based}}
