% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in.R
\name{.filterExp_files}
\alias{.filterExp_files}
\title{filtering Bruker NMR paramter files - helper function read1d}
\usage{
.filterExp_files(pars, exp_type, f_list, n_max)
}
\arguments{
\item{pars}{the parameters}

\item{exp_type}{The experiment type filtered for}

\item{f_list}{list, intact files system for NMR experiments. See fct checkFiles1d}

\item{n_max}{int, maximum number of experiments to read-in}
}
\value{
data frame filtered experiment files
}
\description{
filtering Bruker NMR paramter files - helper function read1d
}
\section{}{
NA
}

\author{
\email{torben.kimhofer@murdoch.edu.au}
}
