% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in.R
\name{extract_pars1d_}
\alias{extract_pars1d_}
\title{Read Bruker NMR paramter files - helper function read1d}
\usage{
extract_pars1d_(f_list)
}
\arguments{
\item{f_list}{list, intact files system for NMR experiments. See fct checkFiles1d}
}
\value{
data frame of spectrometer acquisition metadata
}
\description{
Read Bruker NMR paramter files - helper function read1d
}
\section{}{
NA
}

\author{
\email{torben.kimhofer@murdoch.edu.au}
}
