% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_idx.R
\name{get_idx}
\alias{get_idx}
\title{Get Index}
\usage{
get_idx(sh, ppm)
}
\arguments{
\item{sh}{The lower and upper bounds (2 numeric values) of the chemical shift region you wish to get the index for in ppm}

\item{ppm}{The ppm variable matched to your X}
}
\value{
A numerical array of values that can be used to index both ppm and X to specify a region to visualise or work on
}
\description{
Find the elements of ppm that reside between to points
}
\details{
\code{get_idx()} asks what ppm values are between the points specified in the arg sh and returns this as an array
}
\examples{
# You want to visualise the Creatinine peak in your spectra
data(X, ppm)
idx <- get_idx(sh = c(3, 3.1), ppm = ppm)
plot(ppm[idx], X[1,idx], main = "The Creatinine Peak Found Using get_idx()",
     xlab = "Chemical Shift (ppm)", ylab = "Intensity", type = 'l')
}
\seealso{
Other {estimation}: 
\code{\link{lw}()},
\code{\link{noise}()},
\code{\link{shift_pickr}()}
}
\concept{{estimation}}
