% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmNorm.R
\name{hmNorm}
\alias{hmNorm}
\title{Histogram Matching}
\usage{
hmNorm(
  X,
  noi,
  int_binwid = 0.1,
  alpha = c(0.1, 2),
  use_median = NULL,
  tol = 1e-05
)
}
\arguments{
\item{X}{A numerical matrix containing the NMR spectra to be normalised. Rows should be the spectra and columns being the chemical shift variables}

\item{noi}{The array of maximum noise estimations produced from the function \code{\link[=noise]{noise()}} or provided by the user. Must match X rows.}

\item{int_binwid}{This argument dictates the width of the bins. The average span of intensities is from \code{10}-\code{30}, meaning that an \code{intensity_binwidth} of \code{0.1} would give you 200 bins.}

\item{alpha}{The lower and upper bounds that the golden selection search will search between}

\item{use_median}{This argument dictates whether the function will calculate the median and use that as the reference spectrum or not. If set to \code{FALSE}, the first sample will be used as the reference. This practice is outlined in the methods paper (see 'See also')}

\item{tol}{This defines the tolerance or level of precision the golden selection search will search until. (i.e., it will search until the bounds are \code{tol} distance apart)}
}
\value{
This function assigns the normalised X argument (as X_hm) and the calculated dilution factors (as dilf_hm) to the global environment.
}
\description{
Histogram Matching (HM) is a reference based normalisation that aims to match the histogram created from a experimental spectra with that made from a reference spectra.
}
\details{
\code{hmNorm()} works by selecting range of values to scale X by and then uses a golden selection search to find the value that makes the X overlap the most with the reference spectra.
}
\examples{
# it is mandatory to fill args X, noi, and use_median for `hmNorm()` to work.
data(X, noi)
hmNorm(X, noi, use_median = TRUE, alpha = c(0.1, 2.5))
cat(dilf_hm)
}
\seealso{
\itemize{
\item The methods paper for HM can be found here: \url{http://dx.doi.org/10.1007/s11306-018-1400-6}
\item The paper discussing HM limitations with noise can be found here: \url{http://dx.doi.org/10.1007/s11306-018-1400-6}
}

Other {Reference-Based}: 
\code{\link{pqNorm}()},
\code{\link{q2Norm}()},
\code{\link{xfNorm}()}
}
\author{
\email{kylebario1@gmail.com}
}
\concept{{Reference-Based}}
