% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lw.R
\name{lw}
\alias{lw}
\title{Full Width at Half-Maximum}
\usage{
lw(X, ppm, sh = c(-0.1, 0.1), sf)
}
\arguments{
\item{X}{num matrix, NMR data with rows representing spectra.}

\item{ppm}{num array describing chemical shift positions, its length equals to nrow(X).}

\item{sh}{num array(2), chemical shift range of a singlet for which fwhm is calculated}

\item{sf}{num, operating frequency of the spectrometer (meta$SFO1)}
}
\value{
Array of line widths in ppm. To convert from ppm to Hertz (Hz), multiply values with the spectrometer frequency (column a_SF01 in meta data frame).
}
\description{
Calculating full width at half maximum (FWHM, aka line width). This function returns the ppm difference where peak line crosses half of the peak height. It requires one signal across all spectra within ppm ranges specified in shift.
}
\examples{
read_in(path = system.file('extdata',package='concentr8r'),
        exp_type = list(exp=c("PROF_URINE_NOESY")),
        n_spec='multiple')
lnw <- lw(X, ppm, sh = c(-0.1, 0.1), meta$a_SF01)
bad <- which(lnw>1)
if (length(bad)!=0){cat("The spectra", bad, "have line widths over 1")
} else {cat("All Spectra have line widths under 1")}
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
