% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\docType{data}
\name{meta}
\alias{meta}
\title{Metadata for the NMR spectra.}
\format{
a data frame with 363 variables for two observations
\describe{
\item{DS}{Dummy scans performed prior to acquisition}
\item{NS}{Number of scans performed during the acquisition}
\item{EXP}{The type of experiment/pulse sequence used to acqutire the spectrum}
\item{DATE}{The date at which the sample was finished being acquired}
\item{USER}{Information pertaining to the user as well as other metadata such as the success of the experiment}
}
}
\usage{
meta
}
\description{
A data frame containing acquisition and processing data pertaining to the NMR spectrum.
}
\keyword{datasets}
