% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmo.R
\docType{data}
\name{osmo}
\alias{osmo}
\title{The osmolality of the NMR spectra.}
\format{
Two integer values
}
\usage{
osmo
}
\description{
Two values pertaining the osmolality of the samples.
}
\keyword{datasets}
