% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm.R
\docType{data}
\name{ppm}
\alias{ppm}
\title{The ppm of the NMR spectra.}
\format{
an array with 54692 elements with their value pertaining to the ppm frequency
}
\usage{
ppm
}
\description{
An array of values containing information about frequency of a metabolite signal.
}
\keyword{datasets}
