% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqNorm.R
\name{pqNorm}
\alias{pqNorm}
\title{Probabilistic Quotient Normalisation}
\usage{
pqNorm(
  X,
  noi,
  use_ta = F,
  uv_used = "mode",
  calc_region = c(0.5, 9.5),
  bin_width = 0.01
)
}
\arguments{
\item{X}{The numerical matrix containing the NMR data you wish to normalise. \strong{This should be a preprocessed matrix} with baseline correction, tsp calibration and non-quantitative region removal performed on it. The rows must contain information of one whole spectrum and the columns contain the specific chemical shift variables.}

\item{noi}{Takes an array that is row matched to the X matrix you are normalising with the values equaling the maximum noise estimation for each spectra respectively.}

\item{use_ta}{Requires a boolean \code{TRUE} or \code{FALSE} if total area normalisation should be performed on the spectra before PQN is.}

\item{uv_used}{PQN utilises finding the median or the mode, which are both \emph{U}ni\emph{v}ariate methods. Recognises either the string 'median' or 'mode' to instruct which method to use. Default = 'mode'}

\item{calc_region}{The lower and upper bounds of the spectrum that will be used to calculate the dilution coeficient}

\item{bin_width}{The width of the bin when the spectra are binned}
}
\value{
This function assigns the normalised X argument (as X_pqn) and the calculated dilution factors (as dilf_pqn) to the global environment.
}
\description{
PQN is currently the gold standard method used to normalise NMR spectra.
}
\details{
\code{pqNorm()} functions by calculating the median value for each column and divides each element of a spectrum by its corresponding median. The most frequently occurring quotient (result of division) is then called the dilution factor for that spectrum and all elements within that spectrum are divided by it, scaling the spectrum.
}
\examples{
data(X, noi)
pqNorm(X, noi)
cat(dilf_pqn)
}
\seealso{
The methods paper first describing PQN can be found here: \url{https://doi.org/10.1021/ac051632c}

Other {Reference-Based}: 
\code{\link{hmNorm}()},
\code{\link{q2Norm}()},
\code{\link{xfNorm}()}
}
\author{
\email{kylebario1@gmail.com}
}
\concept{{Reference-Based}}
