% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q1Norm.R
\name{q1Norm}
\alias{q1Norm}
\title{Quantile Normalisation}
\usage{
q1Norm(X)
}
\arguments{
\item{X}{A numerical matrix containing the NMR spectra to be normalised. Rows should be the spectra and columns being the chemical shift variables. Cannot normalise only as single spectrum, X must be a matrix.}
}
\value{
This function assigns the normalised X argument (as X_q1) to the global environment.
}
\description{
A method of NMR spectral normalisation where the maximum intensities
}
\details{
The intensities of each spectrum are ordered smallest to largest, the means of these 'quantiles' are calculated and these means are then reassigned to the ppm they match to (i.e., the ppm which contained the highest intensity will be assigned the mean of the highest values and so on)
}
\examples{
data(X)
q1Norm(X)
}
\seealso{
The methods paper that describes quantile normalisation: \url{https://doi.org/10.1007/s11306-011-0350-z}

Other {Attribute-Based}: 
\code{\link{creNorm}()},
\code{\link{roiNorm}()},
\code{\link{taNorm}()},
\code{\link{vecNorm}()}
}
\author{
\email{kylebario1@gmail.com}
}
\concept{{Attribute-Based}}
