% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q2Norm.R
\name{q2Norm}
\alias{q2Norm}
\title{Quantile/PQN Normalisation}
\usage{
q2Norm(X, uv_used = "mode")
}
\arguments{
\item{X}{A numerical matrix containing the NMR spectra to be normalised. Rows should be the spectra and columns being the chemical shift variables. Cannot normalise only as single spectrum, X must be a matrix.}

\item{uv_used}{The \strong{u}ni\strong{v}ariate measure \strong{used} to calculate the dilution coefficient. \code{mode} or \code{median} are accepted}
}
\value{
This function assigns the normalised X argument (as X_q2) and the calculated dilution factors (as dilf_q2) to the global environment.
}
\description{
A method of NMR spectral normalisation where the maximum intensities
}
\details{
The method orders spectral intensities smallest to largest and finds the mean of each 'quantile' (i.e., the mean of the maximum intensities, the mean of the 2nd maximum intensities). These means are then reassigned back to the ppm that first housed the maximum, 2nd maximum, etc., and PQN is performed on each spectrum using its quantile normalised self. The calculated dilution factor is then used to normalise the original spectrum to reduce manipulation of peak shapes.
}
\examples{
data(X)
q2Norm(X, uv_used = 'mode')
cat(dilf_q2)
}
\seealso{
The methods paper that describes quantile normalisation: \url{https://doi.org/10.1007/s11306-011-0350-z}

Other {Reference-Based}: 
\code{\link{hmNorm}()},
\code{\link{pqNorm}()},
\code{\link{xfNorm}()}
}
\author{
\email{kylebario1@gmail.com}
}
\concept{{Reference-Based}}
