% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roiNorm.R
\name{roiNorm}
\alias{roiNorm}
\title{Region of Interest Normalisation}
\usage{
roiNorm(X, ppm = NULL, sh = c(2.5, 2.75))
}
\arguments{
\item{X}{A numerical matrix with rows being the spectra and columns being the chemical shift variables}

\item{ppm}{A numerical array holding the chemical shift values of the X matrix. Only necessary when X is an array, not when X is a matrix}

\item{sh}{The numerical values defining the lower and upper regions of the Region of Interest. default = c(3,3.1).}
}
\value{
This function assigns the normalised X argument (as X_roi) and the calculated dilution factors (as dilf_roi) to the global environment.
}
\description{
This function normalises the spectra based on a specific area of the spectra. This is helpful when one region is constant across all spectra.
}
\details{
\code{roiNorm()} operates similar to \code{creNorm()} except the normalising factor is not specificly creatinine, it can be what ever region is of interest.
}
\examples{
data(X)
roiNorm(X, sh = c(1.5,2))
cat(dilf_roi)
}
\seealso{
A description of Region of Interest Normalisation can be found in this paper: \url{http://dx.doi.org/10.1007/s11306-018-1400-6}

Other {Attribute-Based}: 
\code{\link{creNorm}()},
\code{\link{q1Norm}()},
\code{\link{taNorm}()},
\code{\link{vecNorm}()}
}
\author{
\email{kylebario1@gmail.com}
}
\concept{{Attribute-Based}}
