% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_pickr.R
\name{shift_pickr}
\alias{shift_pickr}
\title{Peak-Picking}
\usage{
shift_pickr(X, ppm = NULL, sh = c(3, 3.1), pm = 0.005)
}
\arguments{
\item{X}{The X array or matrix you want to pick the peak of}

\item{ppm}{The column matched ppm variable to your X}

\item{sh}{The chemical \strong{sh}ift region you want to search}

\item{pm}{The amount you want to add to each side of the peak}
}
\value{
The ppm variables that match the peak you are searching for
}
\description{
Peak-Picking
}
\examples{
data(X)
idx <- shift_pickr(X[1,], ppm, sh = c(4,4.1), pm = 0.005)
idx1 <- shift_pickr(X[2,], ppm, sh = c(4,4.1), pm = 0.01)
plot(ppm[idx], X[1,idx], main = "Creatinine Peak 4.05", type = 'l', xlim = c(4.05, 4.07))
points(ppm[idx1], X[2,idx1], type = 'l', col = 'red')
}
\seealso{
Other {estimation}: 
\code{\link{noise}()}
}
\concept{{estimation}}
