% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecNorm.R
\name{vecNorm}
\alias{vecNorm}
\title{Vector Length Normalisation}
\usage{
vecNorm(X)
}
\arguments{
\item{X}{The spectr(a/um) to be normalised. Can either be a single spectrum (in the form of an array) or multiple spectra (in the form of a matrix with columns being ppm variables and rows being the spectra)}
}
\value{
This function assigns the normalised X argument (as X_vec) and the calculated dilution factors (as dilf_vec) to the global environment.
}
\description{
This function aims to normalise NMR spectra by their vector length.
}
\details{
This function operates by squaring all values of the spectrum, summing them, and square rooting the result. Then each element from that spectrum is divided by this vector length.
}
\examples{
data(X)
vecNorm(X)
cat(dilf_vec)

}
\seealso{
The methods paper that describes this method of normalisation can be found here: \url{https://doi.org/10.1021/ac051632c}

Other {Attribute-Based}: 
\code{\link{creNorm}()},
\code{\link{q1Norm}()},
\code{\link{roiNorm}()},
\code{\link{taNorm}()}
}
\author{
\email{kylebario1@gmail.com}
}
\concept{{Attribute-Based}}
