% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.R
\docType{data}
\name{X}
\alias{X}
\title{Processed NMR urine spectra.}
\format{
A matrix with 54692 columns which is about the ppm and two rows containing a single spectrum.
}
\usage{
X
}
\description{
An matrix 2 x 54692 of values containing information about metabolites within a urine sample.
}
\keyword{datasets}
