% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfNorm.R
\name{xfNorm}
\alias{xfNorm}
\title{Normalisation based on an External Factor}
\usage{
xfNorm(X, xfactor)
}
\arguments{
\item{X}{A numerical matrix with rows being the spectra and columns being the chemical shift variables}

\item{xfactor}{A numerical array of values that correspond to the dilution coefficients of the spectra. Must be row-matched to the provided X matrix.}
}
\value{
This function returns a list with:
\enumerate{
\item The normalised X matrix in the first list element, and
\item A numerical array of the corresponding dilution factors in the second.
}
\itemize{
\item Following the example below will extract the results quickly and easily.
}
}
\description{
This function utilises an array of dilution coefficients acquired externally to normalise the spectra.
}
\details{
\code{xfNorm()} works by dividing all values within a spectrum by the value of its external factor.
}
\examples{
data(X, osmo)
xfNorm(X, osmo)
}
\seealso{
\url{https://doi.org/10.1021/ac051632c}

Other {Reference-Based}: 
\code{\link{hmNorm}()},
\code{\link{pqNorm}()},
\code{\link{q2Norm}()}
}
\author{
\email{kylebario1@gmail.com}
}
\concept{{Reference-Based}}
